Auto Address Picker from Map View , Place Order UI changes , My Cart UI Changes, Minor issue fixed , Pull To refresh Add in TableView

+++ Actorpay/Model/Items.swift
@@ -15,16 +15,16 @@ public class Items {
     public var description : String?
     public var categoryId : String?
     public var subCategoryId : String?
-    public var actualPrice : Float?
-    public var dealPrice : Float?
+    public var actualPrice : Double?
+    public var dealPrice : Double?
     public var image : String?
     public var merchantId : String?
     public var stockCount : Int?
     public var taxId : String?
     public var stockStatus : String?
     public var status : Bool?
-    public var sgst : Float?
-    public var cgst : Float?
+    public var sgst : Double?
+    public var cgst : Double?
     public var productTaxId : String?
     public var createdAt : String?
     public var updatedAt : String?
@@ -49,16 +49,16 @@ public class Items {
         description = json["description"].string
         categoryId = json["categoryId"].string
         subCategoryId = json["subCategoryId"].string
-        actualPrice = json["actualPrice"].float
-        dealPrice = json["dealPrice"].float
+        actualPrice = json["actualPrice"].double
+        dealPrice = json["dealPrice"].double
         image = json["image"].string
         merchantId = json["merchantId"].string
         stockCount = json["stockCount"].int
         taxId = json["taxId"].string
         stockStatus = json["stockStatus"].string
         status = json["status"].bool
-        sgst = json["sgst"].float
-        cgst = json["cgst"].float
+        sgst = json["sgst"].double
+        cgst = json["cgst"].double
         productTaxId = json["productTaxId"].string
         createdAt = json["createdAt"].string
         updatedAt = json["updatedAt"].string
